Hi there!

First of all thanks for creating this library. We are using it in production on a website with published guidelines on hazardous substances, created in collaboration between ministries and other government bodies in The Netherlands.

Although we've been quite happy with the performance (great job!), our profiling shows that there is some improvement to be made with regards to the creation of temporary objects. This is mostly due to using plain objects instead of ES6 [Maps](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map). 

Objects are great for storing a fixed number of keys, but less great for a larger number of keys, for addition/deletion, or for iterating over them.

In particular the following pattern showed up in a few places:

```
Object.entries(object).forEach(([key, value]) => {
  //
})
```

This has some issues:
* It causes at least one allocation for the array returned by `Object.entries()`.
* It probably causes an allocation for the closure passed into `forEach()`.
* Iterating with `for ... of ... ` is almost always faster (in part due to the allocations).
* Unnecessary temporary objects cause pressure on the garbage collector.

I made the following changes:
* Most occurrences of objects are replaced with Maps.
* Functions that take closures are mostly replaced with loops.
* Some duplicate read/write operations were prevented.
* The internal IDs of fields are now numbers. Because field IDs are consecutive and do not change, we can store the field data in arrays with the ID as the index.
* Document short IDs are now numbers. There is no need to transform them into strings, which saves a small amount of space and time during indexing.

All of this mostly gave a performance boost. Using the benchmarks in the project on my laptop with Node v17.2.0 (which should be representative for a modern version of Chrome):
* Indexing is at least 4x faster.
* Combined search is about 2x faster.

However, **nothing comes for free**... Because the internal structure is no longer 1-1 serialisable to and from JSON, that operation becomes more expensive. Loading an index from JSON is about 30% **slower**, and is also much more complicated. The index format is also slightly different, so the serialised JSON is **not compatible** with the current version (and you might want to consider a major or minor version bump if you accept this PR).

ES6 Maps might also not be supported in very old browsers. However, their support is slightly better than that of Unicode Regexes, which are also used in the project.

Finally, I sacrificed some of your beautiful functional code style on the performance altar.

I'd love to hear your thoughts on this PR. Let me know if there is anything I should clarify!

### Before
```
Index size: 13497 terms, 14097 documents, ~17.49MB in memory, 2.37MB serialized.

Fuzzy search:
=============
  * SearchableMap#fuzzyGet("virtute", 1) x 24,029 ops/sec ±0.23% (100 runs sampled)
  * SearchableMap#fuzzyGet("virtu", 2) x 1,918 ops/sec ±0.23% (99 runs sampled)
  * SearchableMap#fuzzyGet("virtu", 3) x 385 ops/sec ±0.42% (95 runs sampled)

Prefix search:
==============
  * Array.from(SearchableMap#atPrefix("vir")) x 381,810 ops/sec ±0.13% (100 runs sampled)
  * Array.from(SearchableMap#atPrefix("virtut")) x 592,633 ops/sec ±0.16% (98 runs sampled)

Exact search:
=============
  * SearchableMap#get("virtute") x 858,585 ops/sec ±0.08% (99 runs sampled)

Indexing:
=========
  * MiniSearch#addAll(documents) x 3.76 ops/sec ±3.47% (14 runs sampled)

Combined search:
================
  * MiniSearch#search("virtute e conoscienza") x 46.38 ops/sec ±4.34% (63 runs sampled)

Search filtering:
=================
  * MiniSearch#search("virtu", { filter: ... }) x 9,135 ops/sec ±4.19% (88 runs sampled)

Auto suggestion:
================
  * MiniSearch#autoSuggest("virtute cano") x 33,132 ops/sec ±3.56% (91 runs sampled)
  * MiniSearch#autoSuggest("virtue conoscienza", { fuzzy: 0.2 }) x 181,194 ops/sec ±1.20% (101 runs sampled)

Load index:
===========
  * MiniSearch.loadJSON(json, options) x 33.50 ops/sec ±2.46% (60 runs sampled)
```

### After
```
Index size: 13497 terms, 14097 documents, ~16.01MB in memory, 2.49MB serialized.

Fuzzy search:
=============
  * SearchableMap#fuzzyGet("virtute", 1) x 27,119 ops/sec ±0.19% (99 runs sampled)
  * SearchableMap#fuzzyGet("virtu", 2) x 2,253 ops/sec ±0.21% (99 runs sampled)
  * SearchableMap#fuzzyGet("virtu", 3) x 434 ops/sec ±0.38% (92 runs sampled)

Prefix search:
==============
  * Array.from(SearchableMap#atPrefix("vir")) x 446,460 ops/sec ±0.22% (101 runs sampled)
  * Array.from(SearchableMap#atPrefix("virtut")) x 988,842 ops/sec ±0.27% (100 runs sampled)

Exact search:
=============
  * SearchableMap#get("virtute") x 1,823,572 ops/sec ±0.33% (94 runs sampled)

Indexing:
=========
  * MiniSearch#addAll(documents) x 12.09 ops/sec ±2.44% (35 runs sampled)

Combined search:
================
  * MiniSearch#search("virtute e conoscienza") x 91.87 ops/sec ±3.24% (69 runs sampled)

Search filtering:
=================
  * MiniSearch#search("virtu", { filter: ... }) x 18,009 ops/sec ±2.99% (83 runs sampled)

Auto suggestion:
================
  * MiniSearch#autoSuggest("virtute cano") x 58,476 ops/sec ±2.87% (87 runs sampled)
  * MiniSearch#autoSuggest("virtue conoscienza", { fuzzy: 0.2 }) x 327,262 ops/sec ±1.24% (98 runs sampled)

Load index:
===========
  * MiniSearch.loadJSON(json, options) x 23.24 ops/sec ±3.82% (43 runs sampled)
```