@rolftimmermans 2022-02-15
Some things I did not do:
* Measure the impact changing the data structure in the leaf nodes of the radix tree. I think changing the structure to a sparse array could be beneficial.
* Benchmark or improve performance of document removal. It is unlikely to be worse, but it could probably be improved.
* Spend much time on improving the performance of fuzzy search, which is seems to be the largest bottleneck for (our) search performance.