@lucaong 2022-02-16
Hi @rolftimmermans ,
thanks a lot for this outstanding contribution! It makes total sense, and improves performance by a meaningful amount, so I am willing to sacrifice some "functional beauty" as long as the code is still clear.

Allow me some time to review the PR, but I am most likely going to merge it. One question as I approach the review: does turning field IDs to number makes serialized indexes created with a previous version incompatible with this version? I think not, but I only skimmed through the code so far. Asking so I know how to bump the semantic version.

Regarding the slower de-serialization of the index, I think it is a good trade off. As the indexing is now faster, more use-cases can do without the serialization. Moreover, I am planning some breaking changes to the serialization logic in the next major release, which will make this irrelevant.

I am going to elaborate on this, in case you want to provide your opinion. The problems I see with the current serialization logic are:

  - It is very tightly coupled to the internal representation of the index (as you saw in your PR). This makes it fast, but also provides a lot of friction on improving to the internal data structures, as most changes would be breaking changes.
  - Currently the document IDs are mapped to short IDs, and the same happens with the fields. This optimization keeps the serialized index smaller, but I think it's not worth. Avoiding to map the document IDs to short IDs would make the code simpler, possibly improve performance slightly, and the in-memory index might even be smaller, as the strings would be probably pointed to in memory and not duplicated (this should be validated with a benchmark). The JSON index could get larger, because of the repetition of possibly long document IDs, but JSON is usually gzipped by the web server, making this irrelevant for transfer speed.
  - As it is currently implemented, de-serialization cannot be done asynchronously. It would be great to have something similar to `addAllAsync` but for index de-serialization. If such an option was present, a slower de-serialization would be tolerable, as it would not block the UI. In most cases it would also be an improvement for end users, because usually one does not interact with the search immediately, so async de-serialization is likely to be completed by the time the user starts interacting with the search.

Therefore, for the future I am planning to serialize the posting lists (a data structure associating a tokenized term to the list of fields/documents it appears in). The bulk of the serialized structure could at that point be an array of terms to postings, so the deserialization logic could work like `addAllAsync` and leave pauses for the main thread to paint the UI.

On the topic of improving performance of fuzzy search, the best chance I see is to re-implement the [`fuzzySearch`](https://github.com/lucaong/minisearch/blob/9208f91fab9a71c24578da6bf11834a318c342ec/src/SearchableMap/fuzzySearch.ts#L33) to use [this algorithm](http://stevehanov.ca/blog/?id=114) instead of the current one. I suspect that this would get it closer to the theoretical optimum of calculating edit distance on tries, and it is not too hard to implement given the right amount of time.