@rolftimmermans 2022-02-16
Thank you for the detailed response! 

> One question as I approach the review: does turning field IDs to number makes serialized indexes created with a previous version incompatible with this version? I think not, but I only skimmed through the code so far. Asking so I know how to bump the semantic version.

I don't think the change from string to number is a breaking change, because the IDs are still serialised as string keys. However, there **are** breaking changes in the serialisation format in this PR: the field lengths now stored in arrays. This is possible because the IDs are numeric and sequential. This is a win, because arrays are pretty much unbeatable in lookup performance... 

On the topic of numeric IDs: if you keep the numeric short document IDs, it will be possible to change some internal data structures to use arrays instead of Maps. Did I say arrays were great for that? :) I have not gone as far now, but I wouldn't personally get rid of the step from original string ID to short numeric ID without giving this some additional thought and/or testing. (One challenge to address here is that to keep IDs reasonably sequential, removing a document requires some additional bookkeeping.)

To elaborate on this; our app also uses a mapping from original string document ID to (sequential) numeric IDs. This enables us to create [bit vectors](https://en.wikipedia.org/wiki/Bit_array) that indicate the membership of a document in some arbitrary group (like a label, category or tag). Bit vectors are compact and enable extremely fast union/intersection operations. This forms the basis of filtering functionality which complements Minisearch. All of this would be impossible with the original string IDs.