@lucaong 2022-02-16
Thank you @rolftimmermans for your input. While the previous mapping to short string IDs was probably unnecessary, mapping to integers indeed provides interesting options that I did not consider before.

Regarding the bookkeeping upon deletion, I am leaning towards skipping it. On use cases where one adds and deletes a lot of documents, if cleaning up garbage is a concern, one could always re-index the whole collection after a certain number of deletions is reached, and then "hot swap" the MiniSearch instance in use for search. My feeling is that it is worth documenting it, but not to add complex bookkeeping logic.

Now I just need to decide whether to make two major releases (one for this PR, one for further serialization changes), or whether to finally implement the serialization changes and release all in a single major release. In most of my production use-cases I also do not use index serialization, but there are users who rely on that, and there is for example a library out there to serialize the index in Rust, which would be annoying for the author to re-implement twice.

I will immediately release your other PR with a minor version bump though.