@rolftimmermans 2022-02-16
> Now I just need to decide whether to make two major releases (one for this PR, one for further serialization changes), or whether to finally implement the serialization changes and release all in a single major release. In most of my production use-cases I also do not use index serialization, but there are users who rely on that, and there is for example a library out there to serialize the index in Rust, which would be annoying for the author to re-implement twice.

It does make sense to give the serialisation format some more thought and to settle on a reasonable format that decouples it from the internal representation _before_ releasing breaking changes.

An alternative to consider is to spend some effort in removing the backwards incompatibility of the serialisation format in this PR (which means transforming the arrays with field lengths to/from objects). But that's probably only worthwhile if a new serialisation format is unlikely to be worked on by you or someone else soon. What do you think?