@lucaong 2022-02-16
Yeah at the moment I think that serializing an array of postings should be reasonably simple to do. I will give it a try in the next days, and if successful, we can release these changes in a single major version. If it proves to be challenging, I can let you know here and in case work to make this PR backward compatible.